requirements.txt
tensorflow-gpu==1.15.2
numpy==1.18.1
pandas==0.23.4
scipy==1.2.1
